restructure core components
