+++ flightvisualizer-frontend/src/app/core/pipes/int-to-timeofday.pipe.ts
