+++ flightvisualizer-frontend/src/app/core/pipes/pipes.module.ts
