+++ flightvisualizer-frontend/src/app/core/services/data.service.ts
