+++ flightvisualizer-frontend/src/app/core/services/live-feed.service.ts
