+++ flightvisualizer-frontend/src/app/panels/detail-panel/airplane-info/airplane-info.component.ts
@@ -1,5 +1,5 @@
 import {Component, OnDestroy, OnInit} from '@angular/core';
-import {DataStoreService} from "../../../core/service/data-store.service";
+import {DataStoreService} from "../../../core/services/data-store.service";
 import {Subscription} from "rxjs";
 import {LegRender} from "../../../protos/objects";
