+++ flightvisualizer-frontend/src/app/panels/detail-panel/detail-panel.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnDestroy, OnInit} from '@angular/core';
 import {DetailSelectionType, ModeSelection} from "../../core/enum";
 import {Subscription} from "rxjs";
-import {DataStoreService} from "../../core/service/data-store.service";
+import {DataStoreService} from "../../core/services/data-store.service";
 
 @Component({
   selector: 'app-detail-panel',
