+++ flightvisualizer-frontend/src/app/panels/detail-panel/route-info/route-info.component.ts
@@ -1,6 +1,6 @@
 import {AfterViewInit, Component, OnDestroy, OnInit, ViewChild} from '@angular/core';
 import {Subscription} from "rxjs";
-import {DataStoreService} from "../../../core/service/data-store.service";
+import {DataStoreService} from "../../../core/services/data-store.service";
 import {DetailedLegInformation, LegRender} from "../../../protos/objects";
 import {MatPaginator} from "@angular/material/paginator";
 import {MatSort} from "@angular/material/sort";
