+++ flightvisualizer-frontend/src/app/panels/time-panel/live-feed-time-panel/live-feed-time-panel.component.ts
@@ -2,9 +2,9 @@ import { Component, OnDestroy, OnInit } from '@angular/core';
 import { LiveFeedSpeedMultiplier, LiveFeedSpeedMultiplierString } from "../../../core/enum";
 import { Observable, Subscription } from "rxjs";
 import { TimeFilter } from "../../../protos/filters";
-import { DefaultTimeFilter, TimeModifier } from "../../../core/dto/airport";
-import { DataStoreService } from "../../../core/service/data-store.service";
-import { LiveFeedService } from "../../../core/service/live-feed.service";
+import { DefaultTimeFilter } from "../../../core/dto/default-filter";
+import { DataStoreService } from "../../../core/services/data-store.service";
+import { LiveFeedService } from "../../../core/services/live-feed.service";
 
 @Component({
   selector: 'app-live-feed-time-panel',
